adds configurable image interface which makes a difference between existing and new image
